review: change layout - apps internal goes to /internal/%app-name%
