+++ go.mod
@@ -12,6 +12,5 @@ require (
 	github.com/go-chi/chi/v5 v5.2.1
 	github.com/go-resty/resty/v2 v2.16.5
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/spf13/pflag v1.0.6
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
