+++ internal/server/metrics/handlers/errors.go
