+++ internal/server/metrics/handlers/value_handler.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/stepkareserva/obsermon/cmd/server/internal/metrics/service"
+	"github.com/stepkareserva/obsermon/internal/server/metrics/service"
 )
 
 func ValueHandler(s *service.Service) (http.Handler, error) {
